refactor(engine): split engine.rs into boot/execute/init/runtime modules
