+++ packages/engine/src/engine.rs
@@ -7,13 +7,8 @@ use crate::account::{
 use crate::builtin_schema::types::LixVersionDescriptor;
 use crate::builtin_schema::{builtin_schema_definition, builtin_schema_keys};
 use crate::cel::CelEvaluator;
-use crate::deterministic_mode::{
-    load_persisted_sequence_next, load_settings, persist_sequence_highest, DeterministicSettings,
-    RuntimeFunctionProvider,
-};
-use crate::functions::SharedFunctionProvider;
+use crate::deterministic_mode::DeterministicSettings;
 use crate::init::init_backend;
-use crate::json_truthiness::{loosely_false, loosely_true};
 use crate::key_value::{
     key_value_file_id, key_value_plugin_key, key_value_schema_key, key_value_schema_version,
     KEY_VALUE_GLOBAL_VERSION,
@@ -65,7 +60,24 @@ use std::sync::Arc;
 use std::sync::Mutex;
 use std::sync::RwLock;
 
-const DETERMINISTIC_MODE_KEY: &str = "lix_deterministic_mode";
+#[path = "execute/mod.rs"]
+mod execute;
+#[path = "init/active_version.rs"]
+mod init_active_version;
+#[path = "init/bootstrap.rs"]
+mod init_bootstrap;
+#[path = "init/seed.rs"]
+mod init_seed;
+#[path = "plugin/install.rs"]
+mod plugin_install;
+#[path = "runtime_functions.rs"]
+mod runtime_functions;
+
+#[cfg(test)]
+use self::execute::should_sequentialize_postprocess_multi_statement;
+
+pub use crate::boot::{boot, BootAccount, BootArgs, BootKeyValue};
+
 const FILE_DESCRIPTOR_SCHEMA_KEY: &str = "lix_file_descriptor";
 const DIRECTORY_DESCRIPTOR_SCHEMA_KEY: &str = "lix_directory_descriptor";
 #[derive(Debug, Clone, Default)]
@@ -74,39 +86,6 @@ pub struct ExecuteOptions {
 }
 
 pub type EngineTransactionFuture<'a, T> = Pin<Box<dyn Future<Output = Result<T, LixError>> + 'a>>;
-
-#[derive(Debug, Clone)]
-pub struct BootKeyValue {
-    pub key: String,
-    pub value: JsonValue,
-    pub version_id: Option<String>,
-}
-
-#[derive(Debug, Clone)]
-pub struct BootAccount {
-    pub id: String,
-    pub name: String,
-}
-
-pub struct BootArgs {
-    pub backend: Box<dyn LixBackend + Send + Sync>,
-    pub wasm_runtime: Option<Arc<dyn WasmRuntime>>,
-    pub key_values: Vec<BootKeyValue>,
-    pub active_account: Option<BootAccount>,
-    pub access_to_internal: bool,
-}
-
-impl BootArgs {
-    pub fn new(backend: Box<dyn LixBackend + Send + Sync>) -> Self {
-        Self {
-            backend,
-            wasm_runtime: None,
-            key_values: Vec::new(),
-            active_account: None,
-            access_to_internal: false,
-        }
-    }
-}
 
 pub struct Engine {
     backend: Box<dyn LixBackend + Send + Sync>,
@@ -222,7 +201,7 @@ struct TransactionBackendAdapter<'a> {
     _lifetime: PhantomData<&'a ()>,
 }
 
-struct CollectedExecutionSideEffects {
+pub(crate) struct CollectedExecutionSideEffects {
     pending_file_writes: Vec<crate::filesystem::pending_file_writes::PendingFileWrite>,
     pending_file_delete_targets: BTreeSet<(String, String)>,
     detected_file_domain_changes_by_statement: Vec<Vec<DetectedFileDomainChange>>,
@@ -231,7 +210,7 @@ struct CollectedExecutionSideEffects {
 }
 
 #[derive(Default)]
-struct DeferredTransactionSideEffects {
+pub(crate) struct DeferredTransactionSideEffects {
     pending_file_writes: Vec<crate::filesystem::pending_file_writes::PendingFileWrite>,
     pending_file_delete_targets: BTreeSet<(String, String)>,
     detected_file_domain_changes: Vec<DetectedFileDomainChange>,
@@ -288,2617 +267,28 @@ impl<'a> LixBackend for TransactionBackendAdapter<'a> {
     }
 }
 
-pub fn boot(args: BootArgs) -> Engine {
-    let boot_deterministic_settings = infer_boot_deterministic_settings(&args.key_values);
-    let deterministic_boot_pending = boot_deterministic_settings.is_some();
-    Engine {
-        backend: args.backend,
-        wasm_runtime: args.wasm_runtime,
-        cel_evaluator: CelEvaluator::new(),
-        schema_cache: SchemaCache::new(),
-        boot_key_values: args.key_values,
-        boot_active_account: args.active_account,
-        boot_deterministic_settings,
-        deterministic_boot_pending: AtomicBool::new(deterministic_boot_pending),
-        active_version_id: RwLock::new(GLOBAL_VERSION_ID.to_string()),
-        access_to_internal: args.access_to_internal,
-        installed_plugins_cache: RwLock::new(None),
-        plugin_component_cache: Mutex::new(BTreeMap::new()),
-        state_commit_stream_bus: Arc::new(StateCommitStreamBus::default()),
-    }
-}
-
 impl Engine {
-    pub fn wasm_runtime(&self) -> Option<Arc<dyn WasmRuntime>> {
-        self.wasm_runtime.clone()
-    }
-
-    pub fn state_commit_stream(&self, filter: StateCommitStreamFilter) -> StateCommitStream {
-        self.state_commit_stream_bus.subscribe(filter)
-    }
-
-    fn emit_state_commit_stream_changes(&self, changes: Vec<StateCommitStreamChange>) {
-        self.state_commit_stream_bus.emit(changes);
-    }
-
-    async fn load_installed_plugins_with_backend(
-        &self,
-        backend: &dyn LixBackend,
-        allow_cache: bool,
-    ) -> Result<Vec<InstalledPlugin>, LixError> {
-        if allow_cache {
-            let cached = self
-                .installed_plugins_cache
-                .read()
-                .map_err(|_| LixError {
-                    message: "installed plugin cache lock poisoned".to_string(),
-                })?
-                .clone();
-            if let Some(plugins) = cached {
-                return Ok(plugins);
-            }
-        }
-
-        let loaded = crate::plugin::runtime::load_installed_plugins(backend).await?;
-        if allow_cache {
-            let mut guard = self.installed_plugins_cache.write().map_err(|_| LixError {
-                message: "installed plugin cache lock poisoned".to_string(),
-            })?;
-            *guard = Some(loaded.clone());
-        }
-        Ok(loaded)
-    }
-
-    fn invalidate_installed_plugins_cache(&self) -> Result<(), LixError> {
-        let mut guard = self.installed_plugins_cache.write().map_err(|_| LixError {
-            message: "installed plugin cache lock poisoned".to_string(),
-        })?;
-        *guard = None;
-        let mut component_guard = self.plugin_component_cache.lock().map_err(|_| LixError {
-            message: "plugin component cache lock poisoned".to_string(),
-        })?;
-        component_guard.clear();
-        Ok(())
-    }
-
-    pub async fn init(&self) -> Result<(), LixError> {
-        let clear_boot_pending = self.deterministic_boot_pending.load(Ordering::SeqCst);
-        let result = async {
-            init_backend(self.backend.as_ref()).await?;
-            self.ensure_builtin_schemas_installed().await?;
-            let default_active_version_id = self.seed_default_versions().await?;
-            self.seed_default_active_version(&default_active_version_id)
-                .await?;
-            self.seed_default_checkpoint_label().await?;
-            self.seed_boot_key_values().await?;
-            self.seed_boot_account().await?;
-            self.load_and_cache_active_version().await
-        }
-        .await;
-
-        if clear_boot_pending && result.is_ok() {
-            self.deterministic_boot_pending
-                .store(false, Ordering::SeqCst);
-        }
-
-        result
-    }
-
-    pub async fn execute(
-        &self,
-        sql: &str,
-        params: &[Value],
-        options: ExecuteOptions,
-    ) -> Result<QueryResult, LixError> {
-        self.execute_impl(sql, params, options, false).await
-    }
-
-    async fn execute_internal(
-        &self,
-        sql: &str,
-        params: &[Value],
-        options: ExecuteOptions,
-    ) -> Result<QueryResult, LixError> {
-        self.execute_impl(sql, params, options, true).await
-    }
-
-    async fn execute_impl(
-        &self,
-        sql: &str,
-        params: &[Value],
-        options: ExecuteOptions,
-        allow_internal_tables: bool,
-    ) -> Result<QueryResult, LixError> {
-        if !allow_internal_tables && !self.access_to_internal {
-            reject_internal_table_access(sql)?;
-        }
-        let parsed_statements = parse_sql_statements(sql)?;
-        if let Some(statements) =
-            extract_explicit_transaction_script_from_statements(&parsed_statements, params)?
-        {
-            return self
-                .execute_transaction_script_with_options(statements, options)
-                .await;
-        }
-
-        let read_only_query = is_query_only_statements(&parsed_statements);
-        let active_version_id = self.active_version_id.read().unwrap().clone();
-        let writer_key = options.writer_key.as_deref();
-        if read_only_query {
-            self.maybe_refresh_working_change_projection_for_read_query(
-                self.backend.as_ref(),
-                &active_version_id,
-            )
-            .await?;
-        }
-        self.maybe_materialize_reads_with_backend_from_statements(
-            self.backend.as_ref(),
-            &parsed_statements,
-            &active_version_id,
-        )
-        .await?;
-        let should_refresh_file_cache =
-            !read_only_query && should_refresh_file_cache_for_statements(&parsed_statements);
-        let CollectedExecutionSideEffects {
-            pending_file_writes,
-            pending_file_delete_targets,
-            detected_file_domain_changes_by_statement,
-            detected_file_domain_changes,
-            untracked_filesystem_update_domain_changes,
-        } = if read_only_query {
-            CollectedExecutionSideEffects {
-                pending_file_writes: Vec::new(),
-                pending_file_delete_targets: BTreeSet::new(),
-                detected_file_domain_changes_by_statement: Vec::new(),
-                detected_file_domain_changes: Vec::new(),
-                untracked_filesystem_update_domain_changes: Vec::new(),
-            }
-        } else {
-            self.collect_execution_side_effects_with_backend_from_statements(
-                self.backend.as_ref(),
-                &parsed_statements,
-                params,
-                &active_version_id,
-                writer_key,
-                true,
-                true,
-            )
-            .await?
-        };
-        let (settings, sequence_start, functions) = self
-            .prepare_runtime_functions_with_backend(self.backend.as_ref())
-            .await?;
-
-        let output =
-            match preprocess_parsed_statements_with_provider_and_detected_file_domain_changes(
-                self.backend.as_ref(),
-                &self.cel_evaluator,
-                parsed_statements.clone(),
-                params,
-                functions.clone(),
-                &detected_file_domain_changes_by_statement,
-                writer_key,
-            )
-            .await
-            {
-                Ok(output) => output,
-                Err(error)
-                    if should_sequentialize_postprocess_multi_statement_with_statements(
-                        &parsed_statements,
-                        params,
-                        &error,
-                    ) =>
-                {
-                    return self
-                        .execute_multi_statement_sequential_with_options(sql, params, &options)
-                        .await;
-                }
-                Err(error) => return Err(error),
-            };
-        let state_commit_stream_changes =
-            state_commit_stream_changes_from_mutations(&output.mutations, writer_key);
-        let next_active_version_id_from_mutations =
-            active_version_from_mutations(&output.mutations)?;
-        let next_active_version_id_from_updates =
-            active_version_from_update_validations(&output.update_validations)?;
-        if !output.mutations.is_empty() {
-            validate_inserts(self.backend.as_ref(), &self.schema_cache, &output.mutations).await?;
-        }
-        if !output.update_validations.is_empty() {
-            validate_updates(
-                self.backend.as_ref(),
-                &self.schema_cache,
-                &output.update_validations,
-                &output.params,
-            )
-            .await?;
-        }
-        for registration in output.registrations {
-            register_schema(self.backend.as_ref(), &registration.schema_key).await?;
-        }
-        let mut postprocess_file_cache_targets = BTreeSet::new();
-        let mut plugin_changes_committed = false;
-        let result = match output.postprocess {
-            None => {
-                let result = execute_prepared_with_backend(
-                    self.backend.as_ref(),
-                    &output.prepared_statements,
-                )
-                .await?;
-                let tracked_insert_mutation_present = output.mutations.iter().any(|mutation| {
-                    mutation.operation == MutationOperation::Insert && !mutation.untracked
-                });
-                if tracked_insert_mutation_present && !detected_file_domain_changes.is_empty() {
-                    plugin_changes_committed = true;
-                }
-                Ok(result)
-            }
-            Some(postprocess_plan) => {
-                let mut transaction = self.backend.begin_transaction().await?;
-                let result = match execute_prepared_with_transaction(
-                    transaction.as_mut(),
-                    &output.prepared_statements,
-                )
-                .await
-                {
-                    Ok(result) => result,
-                    Err(error) => {
-                        let _ = transaction.rollback().await;
-                        return Err(error);
-                    }
-                };
-                match &postprocess_plan {
-                    PostprocessPlan::VtableUpdate(plan) => {
-                        if should_refresh_file_cache {
-                            let targets = match collect_postprocess_file_cache_targets(
-                                &result.rows,
-                                &plan.schema_key,
-                            ) {
-                                Ok(targets) => targets,
-                                Err(error) => {
-                                    let _ = transaction.rollback().await;
-                                    return Err(error);
-                                }
-                            };
-                            postprocess_file_cache_targets.extend(targets);
-                        }
-                    }
-                    PostprocessPlan::VtableDelete(plan) => {
-                        if should_refresh_file_cache {
-                            let targets = match collect_postprocess_file_cache_targets(
-                                &result.rows,
-                                &plan.schema_key,
-                            ) {
-                                Ok(targets) => targets,
-                                Err(error) => {
-                                    let _ = transaction.rollback().await;
-                                    return Err(error);
-                                }
-                            };
-                            postprocess_file_cache_targets.extend(targets);
-                        }
-                    }
-                }
-                let additional_schema_keys = detected_file_domain_changes
-                    .iter()
-                    .map(|change| change.schema_key.clone())
-                    .collect::<BTreeSet<_>>();
-                for schema_key in additional_schema_keys {
-                    for statement in
-                        register_schema_sql_statements(&schema_key, transaction.dialect())
-                    {
-                        if let Err(error) = transaction.execute(&statement, &[]).await {
-                            let _ = transaction.rollback().await;
-                            return Err(error);
-                        }
-                    }
-                }
-                let mut followup_functions = functions.clone();
-                let followup_statements = match postprocess_plan {
-                    PostprocessPlan::VtableUpdate(plan) => match build_update_followup_sql(
-                        transaction.as_mut(),
-                        &plan,
-                        &result.rows,
-                        &detected_file_domain_changes,
-                        writer_key,
-                        &mut followup_functions,
-                    )
-                    .await
-                    {
-                        Ok(statements) => statements,
-                        Err(error) => {
-                            let _ = transaction.rollback().await;
-                            return Err(error);
-                        }
-                    },
-                    PostprocessPlan::VtableDelete(plan) => match build_delete_followup_sql(
-                        transaction.as_mut(),
-                        &plan,
-                        &result.rows,
-                        &output.params,
-                        &detected_file_domain_changes,
-                        writer_key,
-                        &mut followup_functions,
-                    )
-                    .await
-                    {
-                        Ok(statements) => statements,
-                        Err(error) => {
-                            let _ = transaction.rollback().await;
-                            return Err(error);
-                        }
-                    },
-                };
-                if let Err(error) =
-                    execute_prepared_with_transaction(transaction.as_mut(), &followup_statements)
-                        .await
-                {
-                    let _ = transaction.rollback().await;
-                    return Err(error);
-                }
-                transaction.commit().await?;
-                plugin_changes_committed = true;
-                Ok(result)
-            }
-        }?;
-
-        self.persist_runtime_sequence_with_backend(
-            self.backend.as_ref(),
-            settings,
-            sequence_start,
-            &functions,
-        )
-        .await?;
-
-        if let Some(version_id) =
-            next_active_version_id_from_mutations.or(next_active_version_id_from_updates)
-        {
-            self.set_active_version_id(version_id);
-        }
-        let file_cache_refresh_targets = if should_refresh_file_cache {
-            let mut targets = direct_state_file_cache_refresh_targets(&output.mutations);
-            targets.extend(postprocess_file_cache_targets);
-            targets
-        } else {
-            BTreeSet::new()
-        };
-        let descriptor_cache_eviction_targets =
-            file_descriptor_cache_eviction_targets(&output.mutations);
-        let mut file_cache_invalidation_targets = file_cache_refresh_targets.clone();
-        file_cache_invalidation_targets.extend(descriptor_cache_eviction_targets);
-        file_cache_invalidation_targets.extend(pending_file_delete_targets);
-
-        if !plugin_changes_committed && !detected_file_domain_changes.is_empty() {
-            self.persist_detected_file_domain_changes(&detected_file_domain_changes)
-                .await?;
-        }
-        if !untracked_filesystem_update_domain_changes.is_empty() {
-            self.persist_untracked_file_domain_changes(&untracked_filesystem_update_domain_changes)
-                .await?;
-        }
-        self.persist_pending_file_data_updates(&pending_file_writes)
-            .await?;
-        self.persist_pending_file_path_updates(&pending_file_writes)
-            .await?;
-        self.invalidate_file_data_cache_entries(&file_cache_invalidation_targets)
-            .await?;
-        self.invalidate_file_path_cache_entries(&file_cache_invalidation_targets)
-            .await?;
-        if self.wasm_runtime.is_some() {
-            self.refresh_file_data_for_versions(file_cache_refresh_targets)
-                .await?;
-        }
-        if should_invalidate_installed_plugins_cache_for_statements(&parsed_statements) {
-            self.invalidate_installed_plugins_cache()?;
-        }
-        self.emit_state_commit_stream_changes(state_commit_stream_changes);
-
-        Ok(result)
-    }
-
-    async fn execute_transaction_script_with_options(
-        &self,
-        statements: Vec<Statement>,
-        options: ExecuteOptions,
-    ) -> Result<QueryResult, LixError> {
-        self.execute_statement_script_with_options(statements, &[], &options)
-            .await
-    }
-
-    async fn execute_statement_script_with_options(
-        &self,
-        statements: Vec<Statement>,
-        params: &[Value],
-        options: &ExecuteOptions,
-    ) -> Result<QueryResult, LixError> {
-        let mut transaction = self.backend.begin_transaction().await?;
-        let mut active_version_id = self.active_version_id.read().unwrap().clone();
-        let starting_active_version_id = active_version_id.clone();
-        let mut pending_state_commit_stream_changes = Vec::new();
-        let installed_plugins_cache_invalidation_pending =
-            should_invalidate_installed_plugins_cache_for_statements(&statements);
-        let result = self
-            .execute_statement_script_with_options_in_transaction(
-                transaction.as_mut(),
-                statements,
-                params,
-                options,
-                &mut active_version_id,
-                &mut pending_state_commit_stream_changes,
-            )
-            .await;
-        let result = match result {
-            Ok(result) => result,
-            Err(error) => {
-                let _ = transaction.rollback().await;
-                return Err(error);
-            }
-        };
-
-        transaction.commit().await?;
-        if active_version_id != starting_active_version_id {
-            self.set_active_version_id(active_version_id);
-        }
-        if installed_plugins_cache_invalidation_pending {
-            self.invalidate_installed_plugins_cache()?;
-        }
-        self.emit_state_commit_stream_changes(pending_state_commit_stream_changes);
-        Ok(result)
-    }
-
-    async fn execute_statement_script_with_options_in_transaction(
-        &self,
-        transaction: &mut dyn LixTransaction,
-        original_statements: Vec<Statement>,
-        params: &[Value],
-        options: &ExecuteOptions,
-        active_version_id: &mut String,
-        pending_state_commit_stream_changes: &mut Vec<StateCommitStreamChange>,
-    ) -> Result<QueryResult, LixError> {
-        let coalesced_statements = if params.is_empty() {
-            coalesce_lix_file_transaction_statements(
-                &original_statements,
-                Some(transaction.dialect()),
-            )
-        } else {
-            None
-        };
-        let can_defer_side_effects =
-            params.is_empty() && self.wasm_runtime.is_none() && coalesced_statements.is_some();
-        let mut deferred_side_effects = if can_defer_side_effects {
-            let CollectedExecutionSideEffects {
-                pending_file_writes,
-                pending_file_delete_targets,
-                detected_file_domain_changes: filesystem_tracked_domain_changes,
-                untracked_filesystem_update_domain_changes,
-                ..
-            } = {
-                let backend = TransactionBackendAdapter::new(transaction);
-                self.collect_execution_side_effects_with_backend_from_statements(
-                    &backend,
-                    &original_statements,
-                    params,
-                    active_version_id,
-                    options.writer_key.as_deref(),
-                    false,
-                    false,
-                )
-                .await?
-            };
-            Some(DeferredTransactionSideEffects {
-                pending_file_writes,
-                pending_file_delete_targets: pending_file_delete_targets.clone(),
-                detected_file_domain_changes: filesystem_tracked_domain_changes,
-                untracked_filesystem_update_domain_changes,
-                file_cache_invalidation_targets: pending_file_delete_targets,
-            })
-        } else {
-            None
-        };
-        let sql_statements = if let Some(coalesced) = coalesced_statements {
-            coalesced
-        } else if params.is_empty() {
-            coalesce_vtable_inserts_in_statement_list(original_statements)?
-                .into_iter()
-                .map(|statement| statement.to_string())
-                .collect::<Vec<_>>()
-        } else {
-            original_statements
-                .into_iter()
-                .map(|statement| statement.to_string())
-                .collect::<Vec<_>>()
-        };
-        let skip_statement_side_effect_collection = deferred_side_effects.is_some();
-
-        let mut last_result = QueryResult { rows: Vec::new() };
-        for sql in sql_statements {
-            let result = if skip_statement_side_effect_collection {
-                self.execute_with_options_in_transaction(
-                    transaction,
-                    &sql,
-                    params,
-                    options,
-                    active_version_id,
-                    deferred_side_effects.as_mut(),
-                    true,
-                    pending_state_commit_stream_changes,
-                )
-                .await
-            } else {
-                self.execute_with_options_in_transaction(
-                    transaction,
-                    &sql,
-                    params,
-                    options,
-                    active_version_id,
-                    None,
-                    false,
-                    pending_state_commit_stream_changes,
-                )
-                .await
-            };
-
-            match result {
-                Ok(query_result) => {
-                    last_result = query_result;
-                }
-                Err(error) => {
-                    return Err(error);
-                }
-            }
-        }
-
-        if let Some(side_effects) = deferred_side_effects.as_mut() {
-            self.flush_deferred_transaction_side_effects_in_transaction(
-                transaction,
-                side_effects,
-                options.writer_key.as_deref(),
-            )
-            .await?;
-        }
-        Ok(last_result)
-    }
-
-    async fn begin_transaction_with_options(
-        &self,
-        options: ExecuteOptions,
-    ) -> Result<EngineTransaction<'_>, LixError> {
-        let transaction = self.backend.begin_transaction().await?;
-        Ok(EngineTransaction {
-            engine: self,
-            transaction: Some(transaction),
-            options,
-            active_version_id: self.active_version_id.read().unwrap().clone(),
-            active_version_changed: false,
-            installed_plugins_cache_invalidation_pending: false,
-            pending_state_commit_stream_changes: Vec::new(),
-        })
-    }
-
-    pub async fn transaction<T, F>(&self, options: ExecuteOptions, f: F) -> Result<T, LixError>
-    where
-        F: for<'tx> FnOnce(&'tx mut EngineTransaction<'_>) -> EngineTransactionFuture<'tx, T>,
-    {
-        let mut transaction = self.begin_transaction_with_options(options).await?;
-        match std::panic::AssertUnwindSafe(f(&mut transaction))
-            .catch_unwind()
-            .await
-        {
-            Ok(Ok(value)) => {
-                transaction.commit().await?;
-                Ok(value)
-            }
-            Ok(Err(error)) => {
-                let _ = transaction.rollback().await;
-                Err(error)
-            }
-            Err(payload) => {
-                let _ = transaction.rollback().await;
-                std::panic::resume_unwind(payload);
-            }
-        }
-    }
-
-    async fn execute_with_options_in_transaction(
-        &self,
-        transaction: &mut dyn LixTransaction,
-        sql: &str,
-        params: &[Value],
-        options: &ExecuteOptions,
-        active_version_id: &mut String,
-        deferred_side_effects: Option<&mut DeferredTransactionSideEffects>,
-        skip_side_effect_collection: bool,
-        pending_state_commit_stream_changes: &mut Vec<StateCommitStreamChange>,
-    ) -> Result<QueryResult, LixError> {
-        let parsed_statements = parse_sql_statements(sql)?;
-        let writer_key = options.writer_key.as_deref();
-        let defer_side_effects = deferred_side_effects.is_some();
-        let read_only_query = is_query_only_statements(&parsed_statements);
-        let should_refresh_file_cache =
-            !read_only_query && should_refresh_file_cache_for_statements(&parsed_statements);
-        let (
-            pending_file_writes,
-            pending_file_delete_targets,
-            detected_file_domain_changes,
-            untracked_filesystem_update_domain_changes,
-            settings,
-            sequence_start,
-            functions,
-            output,
-        ) = {
-            let backend = TransactionBackendAdapter::new(transaction);
-            if read_only_query {
-                self.maybe_refresh_working_change_projection_for_read_query(
-                    &backend,
-                    active_version_id.as_str(),
-                )
-                .await?;
-            }
-            self.maybe_materialize_reads_with_backend_from_statements(
-                &backend,
-                &parsed_statements,
-                active_version_id,
-            )
-            .await?;
-            let CollectedExecutionSideEffects {
-                pending_file_writes,
-                pending_file_delete_targets,
-                detected_file_domain_changes_by_statement,
-                detected_file_domain_changes,
-                untracked_filesystem_update_domain_changes,
-            } = if skip_side_effect_collection || read_only_query {
-                CollectedExecutionSideEffects {
-                    pending_file_writes: Vec::new(),
-                    pending_file_delete_targets: BTreeSet::new(),
-                    detected_file_domain_changes_by_statement: Vec::new(),
-                    detected_file_domain_changes: Vec::new(),
-                    untracked_filesystem_update_domain_changes: Vec::new(),
-                }
-            } else {
-                self.collect_execution_side_effects_with_backend_from_statements(
-                    &backend,
-                    &parsed_statements,
-                    params,
-                    active_version_id,
-                    writer_key,
-                    false,
-                    !defer_side_effects,
-                )
-                .await?
-            };
-            let (settings, sequence_start, functions) = self
-                .prepare_runtime_functions_with_backend(&backend)
-                .await?;
-            let output =
-                match preprocess_parsed_statements_with_provider_and_detected_file_domain_changes(
-                    &backend,
-                    &self.cel_evaluator,
-                    parsed_statements.clone(),
-                    params,
-                    functions.clone(),
-                    &detected_file_domain_changes_by_statement,
-                    writer_key,
-                )
-                .await
-                {
-                    Ok(output) => output,
-                    Err(error)
-                        if should_sequentialize_postprocess_multi_statement_with_statements(
-                            &parsed_statements,
-                            params,
-                            &error,
-                        ) =>
-                    {
-                        return Box::pin(
-                            self.execute_multi_statement_sequential_with_options_in_transaction(
-                                transaction,
-                                sql,
-                                params,
-                                options,
-                                active_version_id,
-                                pending_state_commit_stream_changes,
-                            ),
-                        )
-                        .await;
-                    }
-                    Err(error) => return Err(error),
-                };
-            if !output.mutations.is_empty() {
-                validate_inserts(&backend, &self.schema_cache, &output.mutations).await?;
-            }
-            if !output.update_validations.is_empty() {
-                validate_updates(
-                    &backend,
-                    &self.schema_cache,
-                    &output.update_validations,
-                    &output.params,
-                )
-                .await?;
-            }
-
-            (
-                pending_file_writes,
-                pending_file_delete_targets,
-                detected_file_domain_changes,
-                untracked_filesystem_update_domain_changes,
-                settings,
-                sequence_start,
-                functions,
-                output,
-            )
-        };
-        let state_commit_stream_changes =
-            state_commit_stream_changes_from_mutations(&output.mutations, writer_key);
-
-        let next_active_version_id_from_mutations =
-            active_version_from_mutations(&output.mutations)?;
-        let next_active_version_id_from_updates =
-            active_version_from_update_validations(&output.update_validations)?;
-        for registration in output.registrations {
-            for statement in
-                register_schema_sql_statements(&registration.schema_key, transaction.dialect())
-            {
-                transaction.execute(&statement, &[]).await?;
-            }
-        }
-
-        let mut postprocess_file_cache_targets = BTreeSet::new();
-        let mut plugin_changes_committed = false;
-        let result = match output.postprocess {
-            None => {
-                let result =
-                    execute_prepared_with_transaction(transaction, &output.prepared_statements)
-                        .await?;
-                let tracked_insert_mutation_present = output.mutations.iter().any(|mutation| {
-                    mutation.operation == MutationOperation::Insert && !mutation.untracked
-                });
-                if tracked_insert_mutation_present && !detected_file_domain_changes.is_empty() {
-                    plugin_changes_committed = true;
-                }
-                result
-            }
-            Some(postprocess_plan) => {
-                let result =
-                    execute_prepared_with_transaction(transaction, &output.prepared_statements)
-                        .await?;
-                match &postprocess_plan {
-                    PostprocessPlan::VtableUpdate(plan) => {
-                        if should_refresh_file_cache {
-                            postprocess_file_cache_targets.extend(
-                                collect_postprocess_file_cache_targets(
-                                    &result.rows,
-                                    &plan.schema_key,
-                                )?,
-                            );
-                        }
-                    }
-                    PostprocessPlan::VtableDelete(plan) => {
-                        if should_refresh_file_cache {
-                            postprocess_file_cache_targets.extend(
-                                collect_postprocess_file_cache_targets(
-                                    &result.rows,
-                                    &plan.schema_key,
-                                )?,
-                            );
-                        }
-                    }
-                }
-                let additional_schema_keys = detected_file_domain_changes
-                    .iter()
-                    .map(|change| change.schema_key.clone())
-                    .collect::<BTreeSet<_>>();
-                for schema_key in additional_schema_keys {
-                    for statement in
-                        register_schema_sql_statements(&schema_key, transaction.dialect())
-                    {
-                        transaction.execute(&statement, &[]).await?;
-                    }
-                }
-                let mut followup_functions = functions.clone();
-                let followup_statements = match postprocess_plan {
-                    PostprocessPlan::VtableUpdate(plan) => {
-                        build_update_followup_sql(
-                            transaction,
-                            &plan,
-                            &result.rows,
-                            &detected_file_domain_changes,
-                            writer_key,
-                            &mut followup_functions,
-                        )
-                        .await?
-                    }
-                    PostprocessPlan::VtableDelete(plan) => {
-                        build_delete_followup_sql(
-                            transaction,
-                            &plan,
-                            &result.rows,
-                            &output.params,
-                            &detected_file_domain_changes,
-                            writer_key,
-                            &mut followup_functions,
-                        )
-                        .await?
-                    }
-                };
-                execute_prepared_with_transaction(transaction, &followup_statements).await?;
-                plugin_changes_committed = true;
-                result
-            }
-        };
-
-        if let Some(version_id) =
-            next_active_version_id_from_mutations.or(next_active_version_id_from_updates)
-        {
-            *active_version_id = version_id;
-        }
-
-        let file_cache_refresh_targets = if should_refresh_file_cache {
-            let mut targets = direct_state_file_cache_refresh_targets(&output.mutations);
-            targets.extend(postprocess_file_cache_targets);
-            targets
-        } else {
-            BTreeSet::new()
-        };
-        let descriptor_cache_eviction_targets =
-            file_descriptor_cache_eviction_targets(&output.mutations);
-        let mut file_cache_invalidation_targets = file_cache_refresh_targets;
-        file_cache_invalidation_targets.extend(descriptor_cache_eviction_targets);
-        file_cache_invalidation_targets.extend(pending_file_delete_targets);
-
-        if let Some(deferred) = deferred_side_effects {
-            deferred.pending_file_writes.extend(pending_file_writes);
-            deferred
-                .file_cache_invalidation_targets
-                .extend(file_cache_invalidation_targets.iter().cloned());
-            if !plugin_changes_committed {
-                deferred
-                    .detected_file_domain_changes
-                    .extend(detected_file_domain_changes);
-            }
-            deferred
-                .untracked_filesystem_update_domain_changes
-                .extend(untracked_filesystem_update_domain_changes);
-        } else {
-            if !plugin_changes_committed && !detected_file_domain_changes.is_empty() {
-                self.persist_detected_file_domain_changes_in_transaction(
-                    transaction,
-                    &detected_file_domain_changes,
-                )
-                .await?;
-            }
-            if !untracked_filesystem_update_domain_changes.is_empty() {
-                self.persist_untracked_file_domain_changes_in_transaction(
-                    transaction,
-                    &untracked_filesystem_update_domain_changes,
-                )
-                .await?;
-            }
-            self.persist_pending_file_data_updates_in_transaction(
-                transaction,
-                &pending_file_writes,
-            )
-            .await?;
-            self.persist_pending_file_path_updates_in_transaction(
-                transaction,
-                &pending_file_writes,
-            )
-            .await?;
-            self.invalidate_file_data_cache_entries_in_transaction(
-                transaction,
-                &file_cache_invalidation_targets,
-            )
-            .await?;
-            self.invalidate_file_path_cache_entries_in_transaction(
-                transaction,
-                &file_cache_invalidation_targets,
-            )
-            .await?;
-        }
-        self.persist_runtime_sequence_with_backend(
-            &TransactionBackendAdapter::new(transaction),
-            settings,
-            sequence_start,
-            &functions,
-        )
-        .await?;
-
-        pending_state_commit_stream_changes.extend(state_commit_stream_changes);
-        Ok(result)
-    }
-
-    pub async fn install_plugin(
-        &self,
-        manifest_json: &str,
-        wasm_bytes: &[u8],
-    ) -> Result<(), LixError> {
-        let validated = parse_plugin_manifest_json(manifest_json)?;
-        ensure_valid_wasm_binary(wasm_bytes)?;
-        let now = crate::functions::timestamp::timestamp();
-        upsert_plugin_record(
-            self.backend.as_ref(),
-            &validated.manifest,
-            &validated.normalized_json,
-            wasm_bytes,
-            &now,
-        )
-        .await?;
-        self.invalidate_installed_plugins_cache()?;
-        Ok(())
-    }
-
-    pub async fn create_checkpoint(&self) -> Result<crate::CreateCheckpointResult, LixError> {
-        crate::checkpoint::create_checkpoint(self).await
-    }
-
-    /// Exports a portable snapshot as SQLite3 file bytes written via chunk stream.
-    pub async fn export_snapshot(
-        &self,
-        writer: &mut dyn crate::SnapshotChunkWriter,
-    ) -> Result<(), LixError> {
-        self.backend.export_snapshot(writer).await
-    }
-
-    pub async fn restore_from_snapshot(
-        &self,
-        reader: &mut dyn crate::SnapshotChunkReader,
-    ) -> Result<(), LixError> {
-        self.backend.restore_from_snapshot(reader).await?;
-        self.load_and_cache_active_version().await?;
-        self.invalidate_installed_plugins_cache()?;
-        Ok(())
-    }
-
-    pub async fn materialization_plan(
-        &self,
-        req: &MaterializationRequest,
-    ) -> Result<MaterializationPlan, LixError> {
-        crate::materialization::materialization_plan(self.backend.as_ref(), req).await
-    }
-
-    pub async fn apply_materialization_plan(
-        &self,
-        plan: &MaterializationPlan,
-    ) -> Result<MaterializationApplyReport, LixError> {
-        crate::materialization::apply_materialization_plan(self.backend.as_ref(), plan).await
-    }
-
-    pub async fn materialize(
-        &self,
-        req: &MaterializationRequest,
-    ) -> Result<MaterializationReport, LixError> {
-        let plan = crate::materialization::materialization_plan(self.backend.as_ref(), req).await?;
-        let apply =
-            crate::materialization::apply_materialization_plan(self.backend.as_ref(), &plan)
-                .await?;
-
-        if let Some(runtime) = self.wasm_runtime.as_ref() {
-            crate::plugin::runtime::materialize_file_data_with_plugins(
-                self.backend.as_ref(),
-                runtime.as_ref(),
-                &plan,
-            )
-            .await?;
-        }
-
-        Ok(MaterializationReport { plan, apply })
-    }
-
-    async fn maybe_materialize_reads_with_backend_from_statements(
-        &self,
-        backend: &dyn LixBackend,
-        statements: &[Statement],
-        active_version_id: &str,
-    ) -> Result<(), LixError> {
-        let Some(runtime) = self.wasm_runtime.as_ref() else {
-            return Ok(());
-        };
-
-        if let Some(scope) = file_read_materialization_scope_for_statements(statements) {
-            let versions = match scope {
-                FileReadMaterializationScope::ActiveVersionOnly => {
-                    let mut set = BTreeSet::new();
-                    set.insert(active_version_id.to_string());
-                    Some(set)
-                }
-                FileReadMaterializationScope::AllVersions => None,
-            };
-            crate::plugin::runtime::materialize_missing_file_data_with_plugins(
-                backend,
-                runtime.as_ref(),
-                versions.as_ref(),
-            )
-            .await?;
-        }
-        if file_history_read_materialization_required_for_statements(statements) {
-            crate::plugin::runtime::materialize_missing_file_history_data_with_plugins(
-                backend,
-                runtime.as_ref(),
-            )
-            .await?;
-        }
-        Ok(())
-    }
-
-    async fn maybe_refresh_working_change_projection_for_read_query(
-        &self,
-        backend: &dyn LixBackend,
-        active_version_id: &str,
-    ) -> Result<(), LixError> {
-        crate::sql::refresh_working_projection_for_read_query(backend, active_version_id).await
-    }
-
-    async fn collect_execution_side_effects_with_backend_from_statements(
-        &self,
-        backend: &dyn LixBackend,
-        statements: &[Statement],
-        params: &[Value],
-        active_version_id: &str,
-        writer_key: Option<&str>,
-        allow_plugin_cache: bool,
-        detect_plugin_file_changes: bool,
-    ) -> Result<CollectedExecutionSideEffects, LixError> {
-        let pending_file_write_collection =
-            crate::filesystem::pending_file_writes::collect_pending_file_writes_from_statements(
-                backend,
-                statements,
-                params,
-                active_version_id,
-            )
-            .await
-            .map_err(|error| LixError {
-                message: format!("pending file writes collection failed: {}", error.message),
-            })?;
-        let crate::filesystem::pending_file_writes::PendingFileWriteCollection {
-            writes: pending_file_writes,
-            writes_by_statement: pending_file_writes_by_statement,
-        } = pending_file_write_collection;
-        let pending_file_delete_targets =
-            crate::filesystem::pending_file_writes::collect_pending_file_delete_targets_from_statements(
-                backend,
-                statements,
-                params,
-                active_version_id,
-            )
-            .await
-            .map_err(|error| LixError {
-                message: format!("pending file delete collection failed: {}", error.message),
-            })?;
-
-        let detected_file_changes_by_statement = if detect_plugin_file_changes {
-            self.detect_file_changes_for_pending_writes_by_statement_with_backend(
-                backend,
-                &pending_file_writes_by_statement,
-                allow_plugin_cache,
-            )
-            .await?
-        } else {
-            vec![Vec::new(); pending_file_writes_by_statement.len()]
-        };
-        let mut detected_file_domain_changes_by_statement = detected_file_changes_by_statement
-            .into_iter()
-            .map(|changes| {
-                detected_file_domain_changes_from_detected_file_changes(&changes, writer_key)
-            })
-            .collect::<Vec<_>>();
-
-        let filesystem_update_domain_changes =
-            collect_filesystem_update_detected_file_domain_changes_from_statements(
-                backend, statements, params,
-            )
-            .await
-            .map_err(|error| LixError {
-                message: format!(
-                    "filesystem update side-effect detection failed: {}",
-                    error.message
-                ),
-            })?;
-        let filesystem_update_tracked_changes_by_statement = filesystem_update_domain_changes
-            .tracked_changes_by_statement
-            .iter()
-            .map(|changes| detected_file_domain_changes_with_writer_key(changes, writer_key))
-            .collect::<Vec<_>>();
-        let statement_count = detected_file_domain_changes_by_statement
-            .len()
-            .max(filesystem_update_tracked_changes_by_statement.len());
-        detected_file_domain_changes_by_statement.resize_with(statement_count, Vec::new);
-        for (index, tracked_changes) in filesystem_update_tracked_changes_by_statement
-            .into_iter()
-            .enumerate()
-        {
-            detected_file_domain_changes_by_statement[index].extend(tracked_changes);
-            detected_file_domain_changes_by_statement[index] = dedupe_detected_file_domain_changes(
-                &detected_file_domain_changes_by_statement[index],
-            );
-        }
-        let mut detected_file_domain_changes = detected_file_domain_changes_by_statement
-            .iter()
-            .flat_map(|changes| changes.iter().cloned())
-            .collect::<Vec<_>>();
-        detected_file_domain_changes =
-            dedupe_detected_file_domain_changes(&detected_file_domain_changes);
-        let untracked_filesystem_update_domain_changes =
-            dedupe_detected_file_domain_changes(&detected_file_domain_changes_with_writer_key(
-                &filesystem_update_domain_changes.untracked_changes,
-                writer_key,
-            ));
-
-        Ok(CollectedExecutionSideEffects {
-            pending_file_writes,
-            pending_file_delete_targets,
-            detected_file_domain_changes_by_statement,
-            detected_file_domain_changes,
-            untracked_filesystem_update_domain_changes,
-        })
-    }
-
-    async fn flush_deferred_transaction_side_effects_in_transaction(
-        &self,
-        transaction: &mut dyn LixTransaction,
-        side_effects: &mut DeferredTransactionSideEffects,
-        writer_key: Option<&str>,
-    ) -> Result<(), LixError> {
-        let collapsed_writes =
-            collapse_pending_file_writes_for_transaction(&side_effects.pending_file_writes);
-        side_effects.pending_file_writes = collapsed_writes;
-
-        let mut detected_file_domain_changes =
-            std::mem::take(&mut side_effects.detected_file_domain_changes);
-        if !side_effects.pending_file_writes.is_empty() {
-            let detected_by_statement = {
-                let backend = TransactionBackendAdapter::new(transaction);
-                self.detect_file_changes_for_pending_writes_by_statement_with_backend(
-                    &backend,
-                    &[side_effects.pending_file_writes.clone()],
-                    false,
-                )
-                .await?
-            };
-            if let Some(detected) = detected_by_statement.into_iter().next() {
-                detected_file_domain_changes.extend(
-                    detected_file_domain_changes_from_detected_file_changes(&detected, writer_key),
-                );
-            }
-        }
-        detected_file_domain_changes =
-            dedupe_detected_file_domain_changes(&detected_file_domain_changes);
-        let untracked_filesystem_update_domain_changes = dedupe_detected_file_domain_changes(
-            &std::mem::take(&mut side_effects.untracked_filesystem_update_domain_changes),
-        );
-        side_effects
-            .file_cache_invalidation_targets
-            .extend(std::mem::take(
-                &mut side_effects.pending_file_delete_targets,
-            ));
-
-        if !detected_file_domain_changes.is_empty() {
-            self.persist_detected_file_domain_changes_in_transaction(
-                transaction,
-                &detected_file_domain_changes,
-            )
-            .await?;
-        }
-        if !untracked_filesystem_update_domain_changes.is_empty() {
-            self.persist_untracked_file_domain_changes_in_transaction(
-                transaction,
-                &untracked_filesystem_update_domain_changes,
-            )
-            .await?;
-        }
-        self.persist_pending_file_data_updates_in_transaction(
-            transaction,
-            &side_effects.pending_file_writes,
-        )
-        .await?;
-        self.persist_pending_file_path_updates_in_transaction(
-            transaction,
-            &side_effects.pending_file_writes,
-        )
-        .await?;
-        self.invalidate_file_data_cache_entries_in_transaction(
-            transaction,
-            &side_effects.file_cache_invalidation_targets,
-        )
-        .await?;
-        self.invalidate_file_path_cache_entries_in_transaction(
-            transaction,
-            &side_effects.file_cache_invalidation_targets,
-        )
-        .await?;
-
-        Ok(())
-    }
-
-    async fn prepare_runtime_functions_with_backend(
-        &self,
-        backend: &dyn LixBackend,
-    ) -> Result<
-        (
-            DeterministicSettings,
-            i64,
-            SharedFunctionProvider<RuntimeFunctionProvider>,
-        ),
-        LixError,
-    > {
-        let mut settings = load_settings(backend).await?;
-        if self.deterministic_boot_pending.load(Ordering::SeqCst) {
-            if let Some(boot_settings) = self.boot_deterministic_settings {
-                settings = boot_settings;
-            }
-        }
-        let sequence_start = if settings.enabled {
-            load_persisted_sequence_next(backend).await?
-        } else {
-            0
-        };
-        let functions =
-            SharedFunctionProvider::new(RuntimeFunctionProvider::new(settings, sequence_start));
-        Ok((settings, sequence_start, functions))
-    }
-
-    async fn persist_runtime_sequence_with_backend(
-        &self,
-        backend: &dyn LixBackend,
-        settings: DeterministicSettings,
-        sequence_start: i64,
-        functions: &SharedFunctionProvider<RuntimeFunctionProvider>,
-    ) -> Result<(), LixError> {
-        if settings.enabled {
-            let sequence_end = functions.with_lock(|provider| provider.next_sequence());
-            if sequence_end > sequence_start {
-                persist_sequence_highest(backend, sequence_end - 1).await?;
-            }
-        }
-        Ok(())
-    }
-
-    async fn ensure_builtin_schemas_installed(&self) -> Result<(), LixError> {
-        for schema_key in builtin_schema_keys() {
-            let schema = builtin_schema_definition(schema_key).ok_or_else(|| LixError {
-                message: format!("builtin schema '{schema_key}' is not available"),
-            })?;
-            let entity_id = builtin_schema_entity_id(schema)?;
-
-            let existing = self
-                .execute_internal(
-                    "SELECT 1 FROM lix_internal_state_vtable \
-                     WHERE schema_key = 'lix_stored_schema' \
-                       AND entity_id = $1 \
-                       AND version_id = 'global' \
-                       AND snapshot_content IS NOT NULL \
-                     LIMIT 1",
-                    &[Value::Text(entity_id.clone())],
-                    ExecuteOptions::default(),
-                )
-                .await?;
-            if !existing.rows.is_empty() {
-                continue;
-            }
-
-            let snapshot_content = serde_json::json!({
-                "value": schema
-            })
-            .to_string();
-            self.execute_internal(
-                "INSERT INTO lix_internal_state_vtable (schema_key, snapshot_content) \
-                 VALUES ('lix_stored_schema', $1)",
-                &[Value::Text(snapshot_content)],
-                ExecuteOptions::default(),
-            )
-            .await?;
-        }
-
-        Ok(())
-    }
-
-    async fn seed_boot_key_values(&self) -> Result<(), LixError> {
-        for key_value in &self.boot_key_values {
-            let version_id = key_value
-                .version_id
-                .as_deref()
-                .unwrap_or(KEY_VALUE_GLOBAL_VERSION);
-            let snapshot_content = serde_json::json!({
-                "key": key_value.key,
-                "value": key_value.value,
-            })
-            .to_string();
-
-            self.execute_internal(
-                "INSERT INTO lix_internal_state_vtable (\
-                 entity_id, schema_key, file_id, version_id, plugin_key, snapshot_content, schema_version, untracked\
-                 ) VALUES ($1, $2, $3, $4, $5, $6, $7, 1)",
-                &[
-                    Value::Text(key_value.key.clone()),
-                    Value::Text(key_value_schema_key().to_string()),
-                    Value::Text(key_value_file_id().to_string()),
-                    Value::Text(version_id.to_string()),
-                    Value::Text(key_value_plugin_key().to_string()),
-                    Value::Text(snapshot_content),
-                    Value::Text(key_value_schema_version().to_string()),
-                ],
-                ExecuteOptions::default(),
-            )
-            .await?;
-        }
-
-        Ok(())
-    }
-
-    async fn seed_default_checkpoint_label(&self) -> Result<(), LixError> {
-        let existing = self
-            .execute_internal(
-                "SELECT snapshot_content \
-                 FROM lix_internal_state_vtable \
-                 WHERE schema_key = 'lix_label' \
-                   AND file_id = 'lix' \
-                   AND version_id = 'global' \
-                   AND snapshot_content IS NOT NULL",
-                &[],
-                ExecuteOptions::default(),
-            )
-            .await?;
-        for row in &existing.rows {
-            let Some(value) = row.first() else {
-                continue;
-            };
-            let Value::Text(snapshot_content) = value else {
-                continue;
-            };
-            let parsed: JsonValue = serde_json::from_str(snapshot_content).map_err(|error| {
-                LixError {
-                    message: format!("checkpoint label snapshot invalid JSON: {error}"),
-                }
-            })?;
-            if parsed.get("name").and_then(JsonValue::as_str) == Some("checkpoint") {
-                return Ok(());
-            }
-        }
-
-        let label_id = self.generate_runtime_uuid().await?;
-        let snapshot_content = serde_json::json!({
-            "id": label_id,
-            "name": "checkpoint",
-        })
-        .to_string();
-        self.execute_internal(
-            "INSERT INTO lix_internal_state_vtable (\
-             entity_id, schema_key, file_id, version_id, plugin_key, snapshot_content, schema_version, untracked\
-             ) VALUES ($1, 'lix_label', 'lix', 'global', 'lix', $2, '1', 1)",
-            &[Value::Text(label_id), Value::Text(snapshot_content)],
-            ExecuteOptions::default(),
-        )
-        .await?;
-        Ok(())
-    }
-
-    async fn seed_default_versions(&self) -> Result<String, LixError> {
-        self.seed_materialized_version_descriptor(GLOBAL_VERSION_ID, GLOBAL_VERSION_ID, None)
-            .await?;
-        let main_version_id = match self
-            .find_version_id_by_name(DEFAULT_ACTIVE_VERSION_NAME)
-            .await?
-        {
-            Some(version_id) => version_id,
-            None => {
-                let generated_main_id = self.generate_runtime_uuid().await?;
-                self.seed_materialized_version_descriptor(
-                    &generated_main_id,
-                    DEFAULT_ACTIVE_VERSION_NAME,
-                    Some(GLOBAL_VERSION_ID),
-                )
-                .await?;
-                generated_main_id
-            }
-        };
-
-        let bootstrap_commit_id = self
-            .load_latest_commit_id()
-            .await?
-            .unwrap_or_else(|| GLOBAL_VERSION_ID.to_string());
-        if bootstrap_commit_id == GLOBAL_VERSION_ID {
-            self.seed_bootstrap_change_set().await?;
-            self.seed_bootstrap_commit().await?;
+    pub(crate) fn from_boot_args(
+        args: BootArgs,
+        boot_deterministic_settings: Option<DeterministicSettings>,
+    ) -> Self {
+        let deterministic_boot_pending = boot_deterministic_settings.is_some();
+        Self {
+            backend: args.backend,
+            wasm_runtime: args.wasm_runtime,
+            cel_evaluator: CelEvaluator::new(),
+            schema_cache: SchemaCache::new(),
+            boot_key_values: args.key_values,
+            boot_active_account: args.active_account,
+            boot_deterministic_settings,
+            deterministic_boot_pending: AtomicBool::new(deterministic_boot_pending),
+            active_version_id: RwLock::new(GLOBAL_VERSION_ID.to_string()),
+            access_to_internal: args.access_to_internal,
+            installed_plugins_cache: RwLock::new(None),
+            plugin_component_cache: Mutex::new(BTreeMap::new()),
+            state_commit_stream_bus: Arc::new(StateCommitStreamBus::default()),
         }
-        self.seed_materialized_version_pointer(GLOBAL_VERSION_ID, &bootstrap_commit_id)
-            .await?;
-        self.seed_materialized_version_pointer(&main_version_id, &bootstrap_commit_id)
-            .await?;
-
-        Ok(main_version_id)
-    }
-
-    async fn seed_boot_account(&self) -> Result<(), LixError> {
-        let Some(account) = &self.boot_active_account else {
-            return Ok(());
-        };
-
-        let exists = self
-            .execute_internal(
-                "SELECT 1 \
-                 FROM lix_internal_state_vtable \
-                 WHERE schema_key = $1 \
-                   AND entity_id = $2 \
-                   AND file_id = $3 \
-                   AND version_id = $4 \
-                   AND snapshot_content IS NOT NULL \
-                 LIMIT 1",
-                &[
-                    Value::Text(account_schema_key().to_string()),
-                    Value::Text(account.id.clone()),
-                    Value::Text(account_file_id().to_string()),
-                    Value::Text(account_storage_version_id().to_string()),
-                ],
-                ExecuteOptions::default(),
-            )
-            .await?;
-        if exists.rows.is_empty() {
-            self.execute_internal(
-                "INSERT INTO lix_internal_state_vtable (\
-                 entity_id, schema_key, file_id, version_id, plugin_key, snapshot_content, schema_version\
-                 ) VALUES ($1, $2, $3, $4, $5, $6, $7)",
-                &[
-                    Value::Text(account.id.clone()),
-                    Value::Text(account_schema_key().to_string()),
-                    Value::Text(account_file_id().to_string()),
-                    Value::Text(account_storage_version_id().to_string()),
-                    Value::Text(account_plugin_key().to_string()),
-                    Value::Text(account_snapshot_content(&account.id, &account.name)),
-                    Value::Text(account_schema_version().to_string()),
-                ],
-                ExecuteOptions::default(),
-            )
-            .await?;
-        }
-
-        self.execute_internal(
-            "DELETE FROM lix_internal_state_vtable \
-             WHERE untracked = 1 \
-               AND schema_key = $1 \
-               AND file_id = $2 \
-               AND version_id = $3",
-            &[
-                Value::Text(active_account_schema_key().to_string()),
-                Value::Text(active_account_file_id().to_string()),
-                Value::Text(active_account_storage_version_id().to_string()),
-            ],
-            ExecuteOptions::default(),
-        )
-        .await?;
-
-        self.execute_internal(
-            "INSERT INTO lix_internal_state_vtable (\
-             entity_id, schema_key, file_id, version_id, plugin_key, snapshot_content, schema_version, untracked\
-             ) VALUES ($1, $2, $3, $4, $5, $6, $7, 1)",
-            &[
-                Value::Text(account.id.clone()),
-                Value::Text(active_account_schema_key().to_string()),
-                Value::Text(active_account_file_id().to_string()),
-                Value::Text(active_account_storage_version_id().to_string()),
-                Value::Text(active_account_plugin_key().to_string()),
-                Value::Text(active_account_snapshot_content(&account.id)),
-                Value::Text(active_account_schema_version().to_string()),
-            ],
-            ExecuteOptions::default(),
-        )
-        .await?;
-
-        Ok(())
-    }
-
-    async fn seed_materialized_version_descriptor(
-        &self,
-        entity_id: &str,
-        name: &str,
-        inherits_from_version_id: Option<&str>,
-    ) -> Result<(), LixError> {
-        let table = format!(
-            "lix_internal_state_materialized_v1_{}",
-            version_descriptor_schema_key()
-        );
-        let check_sql = format!(
-            "SELECT 1 FROM {table} \
-             WHERE schema_key = '{schema_key}' \
-               AND entity_id = '{entity_id}' \
-               AND file_id = '{file_id}' \
-               AND version_id = '{version_id}' \
-               AND is_tombstone = 0 \
-               AND snapshot_content IS NOT NULL \
-             LIMIT 1",
-            table = table,
-            schema_key = escape_sql_string(version_descriptor_schema_key()),
-            entity_id = escape_sql_string(entity_id),
-            file_id = escape_sql_string(version_descriptor_file_id()),
-            version_id = escape_sql_string(version_descriptor_storage_version_id()),
-        );
-        let existing = self.backend.execute(&check_sql, &[]).await?;
-        if !existing.rows.is_empty() {
-            return Ok(());
-        }
-
-        let snapshot_content =
-            version_descriptor_snapshot_content(entity_id, name, inherits_from_version_id, false);
-        let insert_sql = format!(
-            "INSERT INTO {table} (\
-             entity_id, schema_key, schema_version, file_id, version_id, plugin_key, snapshot_content, \
-             change_id, is_tombstone, created_at, updated_at\
-             ) VALUES (\
-             '{entity_id}', '{schema_key}', '{schema_version}', '{file_id}', '{version_id}', '{plugin_key}', '{snapshot_content}', \
-             'bootstrap', 0, '1970-01-01T00:00:00.000Z', '1970-01-01T00:00:00.000Z'\
-             ) \
-             ON CONFLICT (entity_id, file_id, version_id) DO NOTHING",
-            table = table,
-            entity_id = escape_sql_string(entity_id),
-            schema_key = escape_sql_string(version_descriptor_schema_key()),
-            schema_version = escape_sql_string(version_descriptor_schema_version()),
-            file_id = escape_sql_string(version_descriptor_file_id()),
-            version_id = escape_sql_string(version_descriptor_storage_version_id()),
-            plugin_key = escape_sql_string(version_descriptor_plugin_key()),
-            snapshot_content = escape_sql_string(&snapshot_content),
-        );
-        self.backend.execute(&insert_sql, &[]).await?;
-
-        Ok(())
-    }
-
-    async fn find_version_id_by_name(&self, name: &str) -> Result<Option<String>, LixError> {
-        let table = format!(
-            "lix_internal_state_materialized_v1_{}",
-            version_descriptor_schema_key()
-        );
-        let sql = format!(
-            "SELECT entity_id, snapshot_content \
-             FROM {table} \
-             WHERE schema_key = '{schema_key}' \
-               AND file_id = '{file_id}' \
-               AND version_id = '{version_id}' \
-               AND is_tombstone = 0 \
-               AND snapshot_content IS NOT NULL",
-            table = table,
-            schema_key = escape_sql_string(version_descriptor_schema_key()),
-            file_id = escape_sql_string(version_descriptor_file_id()),
-            version_id = escape_sql_string(version_descriptor_storage_version_id()),
-        );
-        let result = self.backend.execute(&sql, &[]).await?;
-
-        for row in result.rows {
-            if row.len() < 2 {
-                continue;
-            }
-            let entity_id = match &row[0] {
-                Value::Text(value) => value,
-                _ => continue,
-            };
-            let snapshot_content = match &row[1] {
-                Value::Text(value) => value,
-                _ => continue,
-            };
-            let snapshot: LixVersionDescriptor =
-                serde_json::from_str(snapshot_content).map_err(|error| LixError {
-                    message: format!("version descriptor snapshot_content invalid JSON: {error}"),
-                })?;
-            let Some(snapshot_name) = snapshot.name.as_deref() else {
-                continue;
-            };
-            if snapshot_name != name {
-                continue;
-            }
-            let snapshot_id = if snapshot.id.is_empty() {
-                entity_id.as_str()
-            } else {
-                snapshot.id.as_str()
-            };
-            return Ok(Some(snapshot_id.to_string()));
-        }
-
-        Ok(None)
-    }
-
-    async fn seed_materialized_version_pointer(
-        &self,
-        entity_id: &str,
-        commit_id: &str,
-    ) -> Result<(), LixError> {
-        let table = format!(
-            "lix_internal_state_materialized_v1_{}",
-            version_pointer_schema_key()
-        );
-        let check_sql = format!(
-            "SELECT 1 FROM {table} \
-             WHERE schema_key = '{schema_key}' \
-               AND entity_id = '{entity_id}' \
-               AND file_id = '{file_id}' \
-               AND version_id = '{version_id}' \
-               AND is_tombstone = 0 \
-               AND snapshot_content IS NOT NULL \
-             LIMIT 1",
-            table = table,
-            schema_key = escape_sql_string(version_pointer_schema_key()),
-            entity_id = escape_sql_string(entity_id),
-            file_id = escape_sql_string(version_pointer_file_id()),
-            version_id = escape_sql_string(version_pointer_storage_version_id()),
-        );
-        let existing = self.backend.execute(&check_sql, &[]).await?;
-        if !existing.rows.is_empty() {
-            return Ok(());
-        }
-
-        let snapshot_content = version_pointer_snapshot_content(entity_id, commit_id, commit_id);
-        let insert_sql = format!(
-            "INSERT INTO {table} (\
-             entity_id, schema_key, schema_version, file_id, version_id, plugin_key, snapshot_content, \
-             change_id, is_tombstone, created_at, updated_at\
-             ) VALUES (\
-             '{entity_id}', '{schema_key}', '{schema_version}', '{file_id}', '{version_id}', '{plugin_key}', '{snapshot_content}', \
-             'bootstrap', 0, '1970-01-01T00:00:00.000Z', '1970-01-01T00:00:00.000Z'\
-             ) \
-             ON CONFLICT (entity_id, file_id, version_id) DO NOTHING",
-            table = table,
-            entity_id = escape_sql_string(entity_id),
-            schema_key = escape_sql_string(version_pointer_schema_key()),
-            schema_version = escape_sql_string(version_pointer_schema_version()),
-            file_id = escape_sql_string(version_pointer_file_id()),
-            version_id = escape_sql_string(version_pointer_storage_version_id()),
-            plugin_key = escape_sql_string(version_pointer_plugin_key()),
-            snapshot_content = escape_sql_string(&snapshot_content),
-        );
-        self.backend.execute(&insert_sql, &[]).await?;
-
-        Ok(())
-    }
-
-    async fn seed_bootstrap_commit(&self) -> Result<(), LixError> {
-        let existing = self
-            .execute_internal(
-                "SELECT 1 \
-                 FROM lix_internal_state_vtable \
-                 WHERE schema_key = 'lix_commit' \
-                   AND entity_id = $1 \
-                   AND file_id = 'lix' \
-                   AND version_id = 'global' \
-                   AND snapshot_content IS NOT NULL \
-                 LIMIT 1",
-                &[Value::Text(GLOBAL_VERSION_ID.to_string())],
-                ExecuteOptions::default(),
-            )
-            .await?;
-        if !existing.rows.is_empty() {
-            return Ok(());
-        }
-
-        let snapshot_content = serde_json::json!({
-            "id": GLOBAL_VERSION_ID,
-            "change_set_id": GLOBAL_VERSION_ID,
-            "parent_commit_ids": [],
-            "change_ids": [],
-        })
-        .to_string();
-        self.execute_internal(
-            "INSERT INTO lix_internal_state_vtable (\
-             entity_id, schema_key, file_id, version_id, plugin_key, snapshot_content, schema_version, untracked\
-             ) VALUES ($1, 'lix_commit', 'lix', 'global', 'lix', $2, '1', 1)",
-            &[
-                Value::Text(GLOBAL_VERSION_ID.to_string()),
-                Value::Text(snapshot_content),
-            ],
-            ExecuteOptions::default(),
-        )
-        .await?;
-        Ok(())
-    }
-
-    async fn seed_bootstrap_change_set(&self) -> Result<(), LixError> {
-        let existing = self
-            .execute_internal(
-                "SELECT 1 \
-                 FROM lix_internal_state_vtable \
-                 WHERE schema_key = 'lix_change_set' \
-                   AND entity_id = $1 \
-                   AND file_id = 'lix' \
-                   AND version_id = 'global' \
-                   AND snapshot_content IS NOT NULL \
-                 LIMIT 1",
-                &[Value::Text(GLOBAL_VERSION_ID.to_string())],
-                ExecuteOptions::default(),
-            )
-            .await?;
-        if !existing.rows.is_empty() {
-            return Ok(());
-        }
-
-        let snapshot_content = serde_json::json!({
-            "id": GLOBAL_VERSION_ID,
-        })
-        .to_string();
-        self.execute_internal(
-            "INSERT INTO lix_internal_state_vtable (\
-             entity_id, schema_key, file_id, version_id, plugin_key, snapshot_content, schema_version, untracked\
-             ) VALUES ($1, 'lix_change_set', 'lix', 'global', 'lix', $2, '1', 1)",
-            &[
-                Value::Text(GLOBAL_VERSION_ID.to_string()),
-                Value::Text(snapshot_content),
-            ],
-            ExecuteOptions::default(),
-        )
-        .await?;
-        Ok(())
-    }
-
-    async fn seed_default_active_version(&self, version_id: &str) -> Result<(), LixError> {
-        let check_sql = format!(
-            "SELECT 1 \
-             FROM lix_internal_state_untracked \
-             WHERE schema_key = '{schema_key}' \
-               AND file_id = '{file_id}' \
-               AND version_id = '{storage_version_id}' \
-               AND snapshot_content IS NOT NULL \
-             LIMIT 1",
-            schema_key = escape_sql_string(active_version_schema_key()),
-            file_id = escape_sql_string(active_version_file_id()),
-            storage_version_id = escape_sql_string(active_version_storage_version_id()),
-        );
-        let existing = self.backend.execute(&check_sql, &[]).await?;
-        if !existing.rows.is_empty() {
-            return Ok(());
-        }
-
-        let entity_id = self.generate_runtime_uuid().await?;
-        let snapshot_content = active_version_snapshot_content(&entity_id, version_id);
-        let insert_sql = format!(
-            "INSERT INTO lix_internal_state_untracked (\
-             entity_id, schema_key, file_id, version_id, plugin_key, snapshot_content, schema_version, created_at, updated_at\
-             ) VALUES (\
-             '{entity_id}', '{schema_key}', '{file_id}', '{storage_version_id}', '{plugin_key}', '{snapshot_content}', '{schema_version}', '1970-01-01T00:00:00.000Z', '1970-01-01T00:00:00.000Z'\
-             )",
-            entity_id = escape_sql_string(&entity_id),
-            schema_key = escape_sql_string(active_version_schema_key()),
-            file_id = escape_sql_string(active_version_file_id()),
-            storage_version_id = escape_sql_string(active_version_storage_version_id()),
-            plugin_key = escape_sql_string(active_version_plugin_key()),
-            snapshot_content = escape_sql_string(&snapshot_content),
-            schema_version = escape_sql_string(active_version_schema_version()),
-        );
-        self.backend.execute(&insert_sql, &[]).await?;
-        Ok(())
-    }
-
-    async fn load_latest_commit_id(&self) -> Result<Option<String>, LixError> {
-        let result = self
-            .backend
-            .execute(
-                "SELECT entity_id \
-                 FROM lix_internal_change \
-                 WHERE schema_key = 'lix_commit' \
-                 ORDER BY created_at DESC, id DESC \
-                 LIMIT 1",
-                &[],
-            )
-            .await?;
-        let Some(row) = result.rows.first() else {
-            return Ok(None);
-        };
-        let Some(value) = row.first() else {
-            return Ok(None);
-        };
-        match value {
-            Value::Text(value) if !value.is_empty() => Ok(Some(value.clone())),
-            _ => Ok(None),
-        }
-    }
-
-    async fn generate_runtime_uuid(&self) -> Result<String, LixError> {
-        let result = self
-            .execute("SELECT lix_uuid_v7()", &[], ExecuteOptions::default())
-            .await?;
-        let row = result.rows.first().ok_or_else(|| LixError {
-            message: "lix_uuid_v7 query returned no rows".to_string(),
-        })?;
-        let value = row.first().ok_or_else(|| LixError {
-            message: "lix_uuid_v7 query returned no columns".to_string(),
-        })?;
-        match value {
-            Value::Text(text) => Ok(text.clone()),
-            other => Err(LixError {
-                message: format!("lix_uuid_v7 query returned non-text value: {other:?}"),
-            }),
-        }
-    }
-
-    async fn load_and_cache_active_version(&self) -> Result<(), LixError> {
-        let result = self
-            .backend
-            .execute(
-                "SELECT snapshot_content \
-                 FROM lix_internal_state_untracked \
-                 WHERE schema_key = $1 \
-                   AND file_id = $2 \
-                   AND version_id = $3 \
-                   AND snapshot_content IS NOT NULL \
-                 ORDER BY updated_at DESC \
-                 LIMIT 1",
-                &[
-                    Value::Text(active_version_schema_key().to_string()),
-                    Value::Text(active_version_file_id().to_string()),
-                    Value::Text(active_version_storage_version_id().to_string()),
-                ],
-            )
-            .await?;
-
-        if let Some(row) = result.rows.first() {
-            let snapshot_content = row.first().ok_or_else(|| LixError {
-                message: "active version query row is missing snapshot_content".to_string(),
-            })?;
-            let snapshot_content = match snapshot_content {
-                Value::Text(value) => value.as_str(),
-                other => {
-                    return Err(LixError {
-                        message: format!(
-                            "active version snapshot_content must be text, got {other:?}"
-                        ),
-                    })
-                }
-            };
-            let active_version_id = parse_active_version_snapshot(snapshot_content)?;
-            self.set_active_version_id(active_version_id);
-            return Ok(());
-        }
-
-        self.set_active_version_id(GLOBAL_VERSION_ID.to_string());
-        Ok(())
-    }
-
-    async fn detect_file_changes_for_pending_writes_by_statement_with_backend(
-        &self,
-        backend: &dyn LixBackend,
-        writes_by_statement: &[Vec<crate::filesystem::pending_file_writes::PendingFileWrite>],
-        allow_plugin_cache: bool,
-    ) -> Result<Vec<Vec<crate::plugin::runtime::DetectedFileChange>>, LixError> {
-        let Some(runtime) = self.wasm_runtime.as_ref() else {
-            return Ok(vec![Vec::new(); writes_by_statement.len()]);
-        };
-        let installed_plugins = self
-            .load_installed_plugins_with_backend(backend, allow_plugin_cache)
-            .await?;
-        if installed_plugins.is_empty() {
-            return Ok(vec![Vec::new(); writes_by_statement.len()]);
-        }
-
-        let mut loaded_instances = {
-            self.plugin_component_cache
-                .lock()
-                .map_err(|_| LixError {
-                    message: "plugin component cache lock poisoned".to_string(),
-                })?
-                .clone()
-        };
-        let mut detected_by_statement = Vec::with_capacity(writes_by_statement.len());
-        for writes in writes_by_statement {
-            if writes.is_empty() {
-                detected_by_statement.push(Vec::new());
-                continue;
-            }
-
-            let requests = writes
-                .iter()
-                .map(|write| crate::plugin::runtime::FileChangeDetectionRequest {
-                    file_id: write.file_id.clone(),
-                    version_id: write.version_id.clone(),
-                    before_path: write.before_path.clone(),
-                    path: write.path.clone(),
-                    before_data: write.before_data.clone(),
-                    after_data: write.after_data.clone(),
-                })
-                .collect::<Vec<_>>();
-            let detected = crate::plugin::runtime::detect_file_changes_with_plugins_with_cache(
-                backend,
-                runtime.as_ref(),
-                &requests,
-                &installed_plugins,
-                &mut loaded_instances,
-            )
-            .await
-            .map_err(|error| LixError {
-                message: format!("file detect stage failed: {}", error.message),
-            })?;
-            detected_by_statement.push(dedupe_detected_file_changes(&detected));
-        }
-        {
-            let mut guard = self.plugin_component_cache.lock().map_err(|_| LixError {
-                message: "plugin component cache lock poisoned".to_string(),
-            })?;
-            *guard = loaded_instances;
-        }
-
-        Ok(detected_by_statement)
-    }
-
-    async fn persist_detected_file_domain_changes(
-        &self,
-        changes: &[DetectedFileDomainChange],
-    ) -> Result<(), LixError> {
-        self.persist_detected_file_domain_changes_with_untracked(changes, false)
-            .await
-    }
-
-    async fn persist_untracked_file_domain_changes(
-        &self,
-        changes: &[DetectedFileDomainChange],
-    ) -> Result<(), LixError> {
-        self.persist_detected_file_domain_changes_with_untracked(changes, true)
-            .await
-    }
-
-    async fn persist_detected_file_domain_changes_in_transaction(
-        &self,
-        transaction: &mut dyn LixTransaction,
-        changes: &[DetectedFileDomainChange],
-    ) -> Result<(), LixError> {
-        self.persist_detected_file_domain_changes_with_untracked_in_transaction(
-            transaction,
-            changes,
-            false,
-        )
-        .await
-    }
-
-    async fn persist_untracked_file_domain_changes_in_transaction(
-        &self,
-        transaction: &mut dyn LixTransaction,
-        changes: &[DetectedFileDomainChange],
-    ) -> Result<(), LixError> {
-        self.persist_detected_file_domain_changes_with_untracked_in_transaction(
-            transaction,
-            changes,
-            true,
-        )
-        .await
-    }
-
-    async fn persist_detected_file_domain_changes_with_untracked(
-        &self,
-        changes: &[DetectedFileDomainChange],
-        untracked: bool,
-    ) -> Result<(), LixError> {
-        let deduped_changes = dedupe_detected_file_domain_changes(changes);
-        if deduped_changes.is_empty() {
-            return Ok(());
-        }
-
-        let (sql, params) = if untracked {
-            let mut params = Vec::with_capacity(deduped_changes.len() * 10);
-            let mut rows = Vec::with_capacity(deduped_changes.len());
-            for (row_index, change) in deduped_changes.iter().enumerate() {
-                let base = row_index * 10;
-                rows.push(format!(
-                    "(${}, ${}, ${}, ${}, ${}, ${}, ${}, ${}, ${}, ${})",
-                    base + 1,
-                    base + 2,
-                    base + 3,
-                    base + 4,
-                    base + 5,
-                    base + 6,
-                    base + 7,
-                    base + 8,
-                    base + 9,
-                    base + 10
-                ));
-                params.push(Value::Text(change.entity_id.clone()));
-                params.push(Value::Text(change.schema_key.clone()));
-                params.push(Value::Text(change.file_id.clone()));
-                params.push(Value::Text(change.version_id.clone()));
-                params.push(Value::Text(change.plugin_key.clone()));
-                params.push(match &change.snapshot_content {
-                    Some(snapshot_content) => Value::Text(snapshot_content.clone()),
-                    None => Value::Null,
-                });
-                params.push(Value::Text(change.schema_version.clone()));
-                params.push(match &change.metadata {
-                    Some(metadata) => Value::Text(metadata.clone()),
-                    None => Value::Null,
-                });
-                params.push(match &change.writer_key {
-                    Some(writer_key) => Value::Text(writer_key.clone()),
-                    None => Value::Null,
-                });
-                params.push(Value::Integer(1));
-            }
-            (
-                format!(
-                    "INSERT INTO lix_internal_state_vtable (\
-                     entity_id, schema_key, file_id, version_id, plugin_key, snapshot_content, schema_version, metadata, writer_key, untracked\
-                     ) VALUES {}",
-                    rows.join(", ")
-                ),
-                params,
-            )
-        } else {
-            let mut params = Vec::with_capacity(deduped_changes.len() * 9);
-            let mut rows = Vec::with_capacity(deduped_changes.len());
-            for (row_index, change) in deduped_changes.iter().enumerate() {
-                let base = row_index * 9;
-                rows.push(format!(
-                    "(${}, ${}, ${}, ${}, ${}, ${}, ${}, ${}, ${})",
-                    base + 1,
-                    base + 2,
-                    base + 3,
-                    base + 4,
-                    base + 5,
-                    base + 6,
-                    base + 7,
-                    base + 8,
-                    base + 9
-                ));
-                params.push(Value::Text(change.entity_id.clone()));
-                params.push(Value::Text(change.schema_key.clone()));
-                params.push(Value::Text(change.file_id.clone()));
-                params.push(Value::Text(change.version_id.clone()));
-                params.push(Value::Text(change.plugin_key.clone()));
-                params.push(match &change.snapshot_content {
-                    Some(snapshot_content) => Value::Text(snapshot_content.clone()),
-                    None => Value::Null,
-                });
-                params.push(Value::Text(change.schema_version.clone()));
-                params.push(match &change.metadata {
-                    Some(metadata) => Value::Text(metadata.clone()),
-                    None => Value::Null,
-                });
-                params.push(match &change.writer_key {
-                    Some(writer_key) => Value::Text(writer_key.clone()),
-                    None => Value::Null,
-                });
-            }
-            (
-                format!(
-                    "INSERT INTO lix_state_by_version (\
-                     entity_id, schema_key, file_id, version_id, plugin_key, snapshot_content, schema_version, metadata, writer_key\
-                     ) VALUES {}",
-                    rows.join(", ")
-                ),
-                params,
-            )
-        };
-        Box::pin(self.execute_internal(&sql, &params, ExecuteOptions::default())).await?;
-
-        Ok(())
-    }
-
-    async fn persist_detected_file_domain_changes_with_untracked_in_transaction(
-        &self,
-        transaction: &mut dyn LixTransaction,
-        changes: &[DetectedFileDomainChange],
-        untracked: bool,
-    ) -> Result<(), LixError> {
-        let deduped_changes = dedupe_detected_file_domain_changes(changes);
-        if deduped_changes.is_empty() {
-            return Ok(());
-        }
-
-        let (sql, params) = if untracked {
-            let mut params = Vec::with_capacity(deduped_changes.len() * 10);
-            let mut rows = Vec::with_capacity(deduped_changes.len());
-            for (row_index, change) in deduped_changes.iter().enumerate() {
-                let base = row_index * 10;
-                rows.push(format!(
-                    "(${}, ${}, ${}, ${}, ${}, ${}, ${}, ${}, ${}, ${})",
-                    base + 1,
-                    base + 2,
-                    base + 3,
-                    base + 4,
-                    base + 5,
-                    base + 6,
-                    base + 7,
-                    base + 8,
-                    base + 9,
-                    base + 10
-                ));
-                params.push(Value::Text(change.entity_id.clone()));
-                params.push(Value::Text(change.schema_key.clone()));
-                params.push(Value::Text(change.file_id.clone()));
-                params.push(Value::Text(change.version_id.clone()));
-                params.push(Value::Text(change.plugin_key.clone()));
-                params.push(match &change.snapshot_content {
-                    Some(snapshot_content) => Value::Text(snapshot_content.clone()),
-                    None => Value::Null,
-                });
-                params.push(Value::Text(change.schema_version.clone()));
-                params.push(match &change.metadata {
-                    Some(metadata) => Value::Text(metadata.clone()),
-                    None => Value::Null,
-                });
-                params.push(match &change.writer_key {
-                    Some(writer_key) => Value::Text(writer_key.clone()),
-                    None => Value::Null,
-                });
-                params.push(Value::Integer(1));
-            }
-            (
-                format!(
-                    "INSERT INTO lix_internal_state_vtable (\
-                     entity_id, schema_key, file_id, version_id, plugin_key, snapshot_content, schema_version, metadata, writer_key, untracked\
-                     ) VALUES {}",
-                    rows.join(", ")
-                ),
-                params,
-            )
-        } else {
-            let mut params = Vec::with_capacity(deduped_changes.len() * 9);
-            let mut rows = Vec::with_capacity(deduped_changes.len());
-            for (row_index, change) in deduped_changes.iter().enumerate() {
-                let base = row_index * 9;
-                rows.push(format!(
-                    "(${}, ${}, ${}, ${}, ${}, ${}, ${}, ${}, ${})",
-                    base + 1,
-                    base + 2,
-                    base + 3,
-                    base + 4,
-                    base + 5,
-                    base + 6,
-                    base + 7,
-                    base + 8,
-                    base + 9
-                ));
-                params.push(Value::Text(change.entity_id.clone()));
-                params.push(Value::Text(change.schema_key.clone()));
-                params.push(Value::Text(change.file_id.clone()));
-                params.push(Value::Text(change.version_id.clone()));
-                params.push(Value::Text(change.plugin_key.clone()));
-                params.push(match &change.snapshot_content {
-                    Some(snapshot_content) => Value::Text(snapshot_content.clone()),
-                    None => Value::Null,
-                });
-                params.push(Value::Text(change.schema_version.clone()));
-                params.push(match &change.metadata {
-                    Some(metadata) => Value::Text(metadata.clone()),
-                    None => Value::Null,
-                });
-                params.push(match &change.writer_key {
-                    Some(writer_key) => Value::Text(writer_key.clone()),
-                    None => Value::Null,
-                });
-            }
-            (
-                format!(
-                    "INSERT INTO lix_state_by_version (\
-                     entity_id, schema_key, file_id, version_id, plugin_key, snapshot_content, schema_version, metadata, writer_key\
-                     ) VALUES {}",
-                    rows.join(", ")
-                ),
-                params,
-            )
-        };
-        let output = {
-            let backend = TransactionBackendAdapter::new(transaction);
-            preprocess_sql(&backend, &self.cel_evaluator, &sql, &params).await?
-        };
-        execute_prepared_with_transaction(transaction, &output.prepared_statements).await?;
-
-        Ok(())
-    }
-
-    async fn persist_pending_file_data_updates(
-        &self,
-        writes: &[crate::filesystem::pending_file_writes::PendingFileWrite],
-    ) -> Result<(), LixError> {
-        let mut latest_by_key: BTreeMap<(String, String), usize> = BTreeMap::new();
-        for (index, write) in writes.iter().enumerate() {
-            if !write.data_is_authoritative {
-                continue;
-            }
-            latest_by_key.insert((write.file_id.clone(), write.version_id.clone()), index);
-        }
-
-        for index in latest_by_key.into_values() {
-            let write = &writes[index];
-            self.backend
-                .execute(
-                    "INSERT INTO lix_internal_file_data_cache (file_id, version_id, data) \
-                     VALUES ($1, $2, $3) \
-                     ON CONFLICT (file_id, version_id) DO UPDATE SET \
-                     data = EXCLUDED.data",
-                    &[
-                        Value::Text(write.file_id.clone()),
-                        Value::Text(write.version_id.clone()),
-                        Value::Blob(write.after_data.clone()),
-                    ],
-                )
-                .await?;
-        }
-
-        Ok(())
-    }
-
-    async fn persist_pending_file_data_updates_in_transaction(
-        &self,
-        transaction: &mut dyn LixTransaction,
-        writes: &[crate::filesystem::pending_file_writes::PendingFileWrite],
-    ) -> Result<(), LixError> {
-        let mut latest_by_key: BTreeMap<(String, String), usize> = BTreeMap::new();
-        for (index, write) in writes.iter().enumerate() {
-            if !write.data_is_authoritative {
-                continue;
-            }
-            latest_by_key.insert((write.file_id.clone(), write.version_id.clone()), index);
-        }
-
-        for index in latest_by_key.into_values() {
-            let write = &writes[index];
-            transaction
-                .execute(
-                    "INSERT INTO lix_internal_file_data_cache (file_id, version_id, data) \
-                     VALUES ($1, $2, $3) \
-                     ON CONFLICT (file_id, version_id) DO UPDATE SET \
-                     data = EXCLUDED.data",
-                    &[
-                        Value::Text(write.file_id.clone()),
-                        Value::Text(write.version_id.clone()),
-                        Value::Blob(write.after_data.clone()),
-                    ],
-                )
-                .await?;
-        }
-
-        Ok(())
-    }
-
-    async fn persist_pending_file_path_updates(
-        &self,
-        writes: &[crate::filesystem::pending_file_writes::PendingFileWrite],
-    ) -> Result<(), LixError> {
-        let mut latest_by_key: BTreeMap<(String, String), usize> = BTreeMap::new();
-        for (index, write) in writes.iter().enumerate() {
-            latest_by_key.insert((write.file_id.clone(), write.version_id.clone()), index);
-        }
-
-        for index in latest_by_key.into_values() {
-            let write = &writes[index];
-            let Some((name, extension)) = file_name_and_extension_from_path(&write.path) else {
-                continue;
-            };
-            self.backend
-                .execute(
-                    "INSERT INTO lix_internal_file_path_cache \
-                     (file_id, version_id, directory_id, name, extension, path) \
-                     VALUES ($1, $2, NULL, $3, $4, $5) \
-                     ON CONFLICT (file_id, version_id) DO UPDATE SET \
-                     directory_id = EXCLUDED.directory_id, \
-                     name = EXCLUDED.name, \
-                     extension = EXCLUDED.extension, \
-                     path = EXCLUDED.path",
-                    &[
-                        Value::Text(write.file_id.clone()),
-                        Value::Text(write.version_id.clone()),
-                        Value::Text(name),
-                        match extension {
-                            Some(value) => Value::Text(value),
-                            None => Value::Null,
-                        },
-                        Value::Text(write.path.clone()),
-                    ],
-                )
-                .await?;
-        }
-        Ok(())
-    }
-
-    async fn persist_pending_file_path_updates_in_transaction(
-        &self,
-        transaction: &mut dyn LixTransaction,
-        writes: &[crate::filesystem::pending_file_writes::PendingFileWrite],
-    ) -> Result<(), LixError> {
-        let mut latest_by_key: BTreeMap<(String, String), usize> = BTreeMap::new();
-        for (index, write) in writes.iter().enumerate() {
-            latest_by_key.insert((write.file_id.clone(), write.version_id.clone()), index);
-        }
-
-        for index in latest_by_key.into_values() {
-            let write = &writes[index];
-            let Some((name, extension)) = file_name_and_extension_from_path(&write.path) else {
-                continue;
-            };
-            transaction
-                .execute(
-                    "INSERT INTO lix_internal_file_path_cache \
-                     (file_id, version_id, directory_id, name, extension, path) \
-                     VALUES ($1, $2, NULL, $3, $4, $5) \
-                     ON CONFLICT (file_id, version_id) DO UPDATE SET \
-                     directory_id = EXCLUDED.directory_id, \
-                     name = EXCLUDED.name, \
-                     extension = EXCLUDED.extension, \
-                     path = EXCLUDED.path",
-                    &[
-                        Value::Text(write.file_id.clone()),
-                        Value::Text(write.version_id.clone()),
-                        Value::Text(name),
-                        match extension {
-                            Some(value) => Value::Text(value),
-                            None => Value::Null,
-                        },
-                        Value::Text(write.path.clone()),
-                    ],
-                )
-                .await?;
-        }
-        Ok(())
-    }
-
-    async fn invalidate_file_data_cache_entries(
-        &self,
-        targets: &BTreeSet<(String, String)>,
-    ) -> Result<(), LixError> {
-        if targets.is_empty() {
-            return Ok(());
-        }
-
-        const PAIRS_PER_CHUNK: usize = 200;
-        let keys = targets.iter().cloned().collect::<Vec<_>>();
-
-        for chunk in keys.chunks(PAIRS_PER_CHUNK) {
-            let mut params = Vec::with_capacity(chunk.len() * 2);
-            let mut predicates = Vec::with_capacity(chunk.len());
-            for (index, (file_id, version_id)) in chunk.iter().enumerate() {
-                let file_param = index * 2 + 1;
-                let version_param = file_param + 1;
-                predicates.push(format!(
-                    "(file_id = ${file_param} AND version_id = ${version_param})"
-                ));
-                params.push(Value::Text(file_id.clone()));
-                params.push(Value::Text(version_id.clone()));
-            }
-
-            self.backend
-                .execute(
-                    &format!(
-                        "DELETE FROM lix_internal_file_data_cache \
-                         WHERE {}",
-                        predicates.join(" OR ")
-                    ),
-                    &params,
-                )
-                .await?;
-        }
-        Ok(())
-    }
-
-    async fn invalidate_file_data_cache_entries_in_transaction(
-        &self,
-        transaction: &mut dyn LixTransaction,
-        targets: &BTreeSet<(String, String)>,
-    ) -> Result<(), LixError> {
-        if targets.is_empty() {
-            return Ok(());
-        }
-
-        const PAIRS_PER_CHUNK: usize = 200;
-        let keys = targets.iter().cloned().collect::<Vec<_>>();
-
-        for chunk in keys.chunks(PAIRS_PER_CHUNK) {
-            let mut params = Vec::with_capacity(chunk.len() * 2);
-            let mut predicates = Vec::with_capacity(chunk.len());
-            for (index, (file_id, version_id)) in chunk.iter().enumerate() {
-                let file_param = index * 2 + 1;
-                let version_param = file_param + 1;
-                predicates.push(format!(
-                    "(file_id = ${file_param} AND version_id = ${version_param})"
-                ));
-                params.push(Value::Text(file_id.clone()));
-                params.push(Value::Text(version_id.clone()));
-            }
-
-            transaction
-                .execute(
-                    &format!(
-                        "DELETE FROM lix_internal_file_data_cache \
-                         WHERE {}",
-                        predicates.join(" OR ")
-                    ),
-                    &params,
-                )
-                .await?;
-        }
-        Ok(())
-    }
-
-    async fn invalidate_file_path_cache_entries(
-        &self,
-        targets: &BTreeSet<(String, String)>,
-    ) -> Result<(), LixError> {
-        if targets.is_empty() {
-            return Ok(());
-        }
-
-        const PAIRS_PER_CHUNK: usize = 200;
-        let keys = targets.iter().cloned().collect::<Vec<_>>();
-
-        for chunk in keys.chunks(PAIRS_PER_CHUNK) {
-            let mut params = Vec::with_capacity(chunk.len() * 2);
-            let mut predicates = Vec::with_capacity(chunk.len());
-            for (index, (file_id, version_id)) in chunk.iter().enumerate() {
-                let file_param = index * 2 + 1;
-                let version_param = file_param + 1;
-                predicates.push(format!(
-                    "(file_id = ${file_param} AND version_id = ${version_param})"
-                ));
-                params.push(Value::Text(file_id.clone()));
-                params.push(Value::Text(version_id.clone()));
-            }
-
-            self.backend
-                .execute(
-                    &format!(
-                        "DELETE FROM lix_internal_file_path_cache \
-                         WHERE {}",
-                        predicates.join(" OR ")
-                    ),
-                    &params,
-                )
-                .await?;
-        }
-        Ok(())
-    }
-
-    async fn invalidate_file_path_cache_entries_in_transaction(
-        &self,
-        transaction: &mut dyn LixTransaction,
-        targets: &BTreeSet<(String, String)>,
-    ) -> Result<(), LixError> {
-        if targets.is_empty() {
-            return Ok(());
-        }
-
-        const PAIRS_PER_CHUNK: usize = 200;
-        let keys = targets.iter().cloned().collect::<Vec<_>>();
-
-        for chunk in keys.chunks(PAIRS_PER_CHUNK) {
-            let mut params = Vec::with_capacity(chunk.len() * 2);
-            let mut predicates = Vec::with_capacity(chunk.len());
-            for (index, (file_id, version_id)) in chunk.iter().enumerate() {
-                let file_param = index * 2 + 1;
-                let version_param = file_param + 1;
-                predicates.push(format!(
-                    "(file_id = ${file_param} AND version_id = ${version_param})"
-                ));
-                params.push(Value::Text(file_id.clone()));
-                params.push(Value::Text(version_id.clone()));
-            }
-
-            transaction
-                .execute(
-                    &format!(
-                        "DELETE FROM lix_internal_file_path_cache \
-                         WHERE {}",
-                        predicates.join(" OR ")
-                    ),
-                    &params,
-                )
-                .await?;
-        }
-        Ok(())
-    }
-
-    async fn refresh_file_data_for_versions(
-        &self,
-        targets: BTreeSet<(String, String)>,
-    ) -> Result<(), LixError> {
-        let versions = targets
-            .into_iter()
-            .map(|(_, version_id)| version_id)
-            .collect::<BTreeSet<_>>();
-        if versions.is_empty() {
-            return Ok(());
-        }
-
-        self.materialize(&MaterializationRequest {
-            scope: MaterializationScope::Versions(versions),
-            debug: MaterializationDebugMode::Off,
-            debug_row_limit: 1,
-        })
-        .await?;
-        Ok(())
-    }
-
-    fn set_active_version_id(&self, version_id: String) {
-        let mut guard = self.active_version_id.write().unwrap();
-        if *guard == version_id {
-            return;
-        }
-        *guard = version_id;
-    }
-
-    async fn execute_multi_statement_sequential_with_options(
-        &self,
-        sql: &str,
-        params: &[Value],
-        options: &ExecuteOptions,
-    ) -> Result<QueryResult, LixError> {
-        let statements = parse_sql_statements(sql)?;
-        self.execute_statement_script_with_options(statements, params, options)
-            .await
-    }
-
-    async fn execute_multi_statement_sequential_with_options_in_transaction(
-        &self,
-        transaction: &mut dyn LixTransaction,
-        sql: &str,
-        params: &[Value],
-        options: &ExecuteOptions,
-        active_version_id: &mut String,
-        pending_state_commit_stream_changes: &mut Vec<StateCommitStreamChange>,
-    ) -> Result<QueryResult, LixError> {
-        let statements = parse_sql_statements(sql)?;
-        self.execute_statement_script_with_options_in_transaction(
-            transaction,
-            statements,
-            params,
-            options,
-            active_version_id,
-            pending_state_commit_stream_changes,
-        )
-        .await
-    }
-}
-
-async fn execute_prepared_with_backend(
-    backend: &dyn LixBackend,
-    statements: &[crate::sql::PreparedStatement],
-) -> Result<QueryResult, LixError> {
-    let mut last_result = QueryResult { rows: Vec::new() };
-    for statement in statements {
-        last_result = backend.execute(&statement.sql, &statement.params).await?;
-    }
-    Ok(last_result)
-}
-
-async fn execute_prepared_with_transaction(
-    transaction: &mut dyn LixTransaction,
-    statements: &[crate::sql::PreparedStatement],
-) -> Result<QueryResult, LixError> {
-    let mut last_result = QueryResult { rows: Vec::new() };
-    for statement in statements {
-        last_result = transaction
-            .execute(&statement.sql, &statement.params)
-            .await?;
-    }
-    Ok(last_result)
-}
-
-#[cfg(test)]
-fn should_sequentialize_postprocess_multi_statement(
-    sql: &str,
-    params: &[Value],
-    error: &LixError,
-) -> bool {
-    let Ok(statements) = parse_sql_statements(sql) else {
-        return false;
-    };
-    should_sequentialize_postprocess_multi_statement_with_statements(&statements, params, error)
-}
-
-fn should_sequentialize_postprocess_multi_statement_with_statements(
-    statements: &[Statement],
-    params: &[Value],
-    error: &LixError,
-) -> bool {
-    if !params.is_empty() || !is_postprocess_multi_statement_error(&error.message) {
-        return false;
-    }
-    if statements.len() <= 1 {
-        return false;
     }
-
-    !statements.iter().any(|statement| {
-        matches!(
-            statement,
-            Statement::StartTransaction { .. }
-                | Statement::Commit { .. }
-                | Statement::Rollback { .. }
-        )
-    })
-}
-
-fn is_postprocess_multi_statement_error(message: &str) -> bool {
-    matches!(
-        message,
-        "postprocess rewrites require a single statement"
-            | "only one postprocess rewrite is supported per query"
-    )
 }
 
 fn file_name_and_extension_from_path(path: &str) -> Option<(String, Option<String>)> {
@@ -4171,83 +1561,6 @@ fn builtin_schema_entity_id(schema: &JsonValue) -> Result<String, LixError> {
         })?;
 
     Ok(format!("{schema_key}~{schema_version}"))
-}
-
-async fn upsert_plugin_record(
-    backend: &dyn LixBackend,
-    manifest: &PluginManifest,
-    manifest_json: &str,
-    wasm_bytes: &[u8],
-    timestamp: &str,
-) -> Result<(), LixError> {
-    backend
-        .execute(
-            "INSERT INTO lix_internal_plugin (\
-             key, runtime, api_version, detect_changes_glob, entry, manifest_json, wasm, created_at, updated_at\
-             ) VALUES ($1, $2, $3, $4, $5, $6, $7, $8, $8) \
-             ON CONFLICT (key) DO UPDATE SET \
-             runtime = EXCLUDED.runtime, \
-             api_version = EXCLUDED.api_version, \
-             detect_changes_glob = EXCLUDED.detect_changes_glob, \
-             entry = EXCLUDED.entry, \
-             manifest_json = EXCLUDED.manifest_json, \
-             wasm = EXCLUDED.wasm, \
-             updated_at = EXCLUDED.updated_at",
-            &[
-                Value::Text(manifest.key.clone()),
-                Value::Text(manifest.runtime.as_str().to_string()),
-                Value::Text(manifest.api_version.clone()),
-                Value::Text(manifest.file_match.path_glob.clone()),
-                Value::Text(manifest.entry_or_default().to_string()),
-                Value::Text(manifest_json.to_string()),
-                Value::Blob(wasm_bytes.to_vec()),
-                Value::Text(timestamp.to_string()),
-            ],
-        )
-        .await?;
-
-    Ok(())
-}
-
-fn ensure_valid_wasm_binary(wasm_bytes: &[u8]) -> Result<(), LixError> {
-    if wasm_bytes.is_empty() {
-        return Err(LixError {
-            message: "Plugin wasm bytes must not be empty".to_string(),
-        });
-    }
-    if wasm_bytes.len() < 8 || !wasm_bytes.starts_with(&[0x00, 0x61, 0x73, 0x6d]) {
-        return Err(LixError {
-            message: "Plugin wasm bytes must start with a valid wasm header".to_string(),
-        });
-    }
-    Ok(())
-}
-
-fn infer_boot_deterministic_settings(key_values: &[BootKeyValue]) -> Option<DeterministicSettings> {
-    key_values.iter().rev().find_map(|key_value| {
-        if key_value.key != DETERMINISTIC_MODE_KEY {
-            return None;
-        }
-        if key_value
-            .version_id
-            .as_deref()
-            .is_some_and(|version| version != KEY_VALUE_GLOBAL_VERSION)
-        {
-            return None;
-        }
-        let object = key_value.value.as_object()?;
-        let enabled = object.get("enabled").map(loosely_true).unwrap_or(false);
-        if !enabled {
-            return None;
-        }
-        let uuid_v7_enabled = !object.get("uuid_v7").map(loosely_false).unwrap_or(false);
-        let timestamp_enabled = !object.get("timestamp").map(loosely_false).unwrap_or(false);
-        Some(DeterministicSettings {
-            enabled,
-            uuid_v7_enabled,
-            timestamp_enabled,
-        })
-    })
 }
 
 fn active_version_from_mutations(mutations: &[MutationRow]) -> Result<Option<String>, LixError> {
